Addressed first round of code reviews
* reformatted and consolidated TODOs
* fixed typos and docstrings
* created StrEnum to be used as base class for all string enums
* renamed whylogs_utils.py to whylogs_context.py
