+++ src/zenml/integrations/whylogs/materializers/whylogs_materializer.py
@@ -26,7 +26,7 @@ PROFILE_FILENAME = "profile.pb"
 
 
 class WhylogsMaterializer(BaseMaterializer):
-    """Materializer to read/write Pytorch models."""
+    """Materializer to read/write whylogs dataset profiles."""
 
     ASSOCIATED_TYPES = [DatasetProfile]
     ASSOCIATED_ARTIFACT_TYPES = [StatisticsArtifact]
@@ -55,6 +55,8 @@ class WhylogsMaterializer(BaseMaterializer):
         with fileio.open(filepath, "wb") as f:
             f.write(protobuf)
 
-        # TODO: find another way to detect if whylabs is enabled
+        # TODO [LOW]: uploading profiles to whylabs should be enabled and
+        #  configurable at step level or pipeline level instead of being
+        #  globally enabled.
         if os.environ.get("WHYLABS_DEFAULT_ORG_ID"):
             upload_profile(profile)
